% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substitute.R
\name{function.substitute}
\alias{function.substitute}
\title{Substitute Parameters with Values}
\usage{
function.substitute(f, substitutes)
}
\arguments{
\item{f}{the function}

\item{substitutes}{the arguments to be substituted into the function.}
}
\description{
Create a copy of a given function where a set of parameters are
  replaced with values.
}
\examples{
f <- function(x, pars) pars[1] + pars[2]*(x + pars[3]*x)
f
# function(x, pars) pars[1] + pars[2]*(x + pars[3]*x)
 g <- function.substitute(f, list("pars"=c(1, 2, 3)))
 g
# function (x)
#  1 + 2 * (x + 3 * x)
h <- function.substitute(f, list("pars"=c(1, 2, 3), "x"=4))
# function ()
# 33
}
